+++ src/Sidebar/Searchbar/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 
 
-import '.style.css'
+import './style.css'
 
 
 export default ()=> {
