+++ src/first page/Card/index.js
@@ -4,14 +4,14 @@ import './style.css'
 
 import Itemimg from './images/bitmap1.jpg'
 
-import Sale_label from '../SaleLabel'
+import SaleLabel from '../../SaleLabel'
 
 export default ()=> {
   return(
     <div className="col-4">
-      <Sale_label />
+      <SaleLabel />
       <img className="item-image" src={Itemimg} alt="item-img" />
-      <a href="#" className="item-price-sale">$170</a>
+      <a href="" className="item-price-sale">$170</a>
     </div>
   );
 };
