+++ src/style.css
@@ -1,3 +1,9 @@
+body{
+  margin: 0;
+  padding:0;
+  display: flex;
+}
+
 .shoes{
   width: 1025px;
 }
